Set Shapefile column names given dictionary names fetched from .xlsx data dictionary file.

Set Shapefile column names given dictionary names fetched from .xlsx data dictionary file.

get_geomac_new_cnames(fpath, year, ddict_type, ddict_xlsx_path, cols_rng)

Arguments

fpath
: shapefile path

year
: year the shapefile is referred to

ddict_type
: dictionary type - this needs to be corresponding to one of the data dictionaries retrieved in the .xlsx

ddict_xlsx_path
: path of the data dictionary .xlsx

cols_rng
: column ranges to extract from the .xlsx

Value

Shapefile dataframe