Wrapper for NOOA STORMEVENTS CSV data extraction pipeline. This function will setup global variables, gather information and urls on data and metadata and download the data.

Wrapper for NOOA STORMEVENTS CSV data extraction pipeline. This function will setup global variables, gather information and urls on data and metadata and download the data.

noaa_stormevents_extract(year_periods = 1950:1950,
  data_folder = "data", dat_flag = "noaa_stormevents",
  remove = FALSE, data_dl_method = "libcurl", url = NULL,
  metadata_dirname = "metadata")

Arguments

year_periods
(numeric) : 1950:1950 (default) array of years to be downloaded

data_folder
(character) : "data" (default) data folder name

dat_flag
(character) : "noaa_stormevents" (default) name of the data - will appear as data folder name when data are saved

remove
(logical) : FALSE (default) remove .zip file after extraction

data_dl_method
(character) : "libcurl" (default) data download method

url
(character) : NULL (default) data url - the default one is encoded in the function body

metadata_dirname
(character) : "metadata" (default) metadata folder name

Examples

if (FALSE) {
noaa_stormevents_extract(year_periods = 1800:1801)
}