Extracts the TIGER/Line Files and Shapefiles for US States, US Counties, and US military lines for a given year, and resolution, in preparation for transform purposes

Extracts the TIGER/Line Files and Shapefiles for US States, US Counties, and US military lines for a given year, and resolution, in preparation for transform purposes

tigris_extract(tig_res = "5m", tig_year = 2019, tig_cb = FALSE, tig_crs = 4326)

Arguments

tig_res
(character) : The resolution of the cartographic boundary file (if cb == TRUE). Defaults to '500k'; options include '5m' (1:5 million) and '20m' (1:20 million). Default is '5m'

tig_year
(integer) : The year for which you want to download the

tig_cb
(logical) : If cb is set to TRUE, download a generalized (1:500k) states file. Defaults to FALSE (the most detailed TIGER/Line file)

tig_crs
(integer) : The Coordinate Reference System (CRS) system with default value of 4326

Value

(list) : A list of shapefiles with cleaned up column names for US States, US Counties, and US military lines for a given year, and resolution, with cleaned column names, ready for modeling transformations.

Examples

if (FALSE) {
tigris_out <- tigris_extract(tig_res = '5m', tig_year = 2019,
                            tig_cb = FALSE, tig_crs = 4326)
}